requirements.txt
pandas==1.3.3
yfinance==0.1.63
google-auth-oauthlib==0.4.6
google-auth-httplib2==0.1.0
google-api-python-client==2.23.0